Skul rizki2:
1.titipan kevin tingkatan kelas 
2. absensi 
3. siswa japel view

+++ app/Http/Controllers/Admin/Referensi/TingkatanKelasController.php
@@ -13,7 +13,7 @@ class TingkatanKelasController extends Controller
 {
     public function index(Request $request) {
         if ($request->ajax()) {
-            $data = TingkatanKelas::latest()->get();
+            $data = TingkatanKelas::where('sekolah_id', auth()->user()->id_sekolah)->latest()->get();
             return DataTables::of($data)
                 ->addColumn('action', function ($data) {
                     $button = '<button type="button" id="'.$data->id.'" class="edit btn btn-mini btn-info shadow-sm">Edit</button>';
@@ -25,7 +25,8 @@ class TingkatanKelasController extends Controller
                 ->make(true);
         }
         
-        return view('admin.referensi.tingkatan-kelas', ['mySekolah' => User::sekolah()]);
+        // , ['mySekolah' => User::sekolah()]
+        return view('admin.referensi.tingkatan-kelas');
     }
 
     public function store(Request $request) {
@@ -48,6 +49,7 @@ class TingkatanKelasController extends Controller
         }
 
         $status = TingkatanKelas::create([
+            'sekolah_id' => auth()->user()->id_sekolah,
             'name'  => $request->input('tingkat'),
         ]);
 
@@ -84,10 +86,12 @@ class TingkatanKelasController extends Controller
                     'errors' => $validator->errors()->all()
                 ]);
         }
-
-        $status = TingkatanKelas::whereId($request->input('hidden_id'))->update([
-            'name'  => $request->input('tingkat'),
-        ]);
+        $status = TingkatanKelas::where([
+                    ['id', $request->input('hidden_id')],
+                    ['sekolah_id', auth()->user()->id_sekolah]
+                ])->update([
+                    'name'  => $request->input('tingkat'),
+                ]);
 
         return response()
             ->json([
@@ -97,6 +101,10 @@ class TingkatanKelasController extends Controller
 
     public function destroy($id) {
         $tingkat = TingkatanKelas::find($id);
+        if ($tingkat->sekolah_id != auth()->user()->id_sekolah) {
+            return;
+        }
+
         $tingkat->delete();
     }
 }
